solution for issue no. 1
https://github.com/Crunch-io/lazycsv/issues/1

Fixes issue reported by @krzysztofspalinski where files relative to the cwd are not found. This was due to lazycsv preferentially locating files relative to the __file__ global parameter, and searching only relative to __file__. Now, lazycsv first checks to see if the file is relative to cwd, and if not only then does it check relative to __file__, if __file__ exists.